ai-quiz-generator/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
langchain>=0.3.4,<0.4
langchain-core>=0.3.15,<0.4
langchain-community>=0.3.12,<0.4
langchain-google-genai>=2.0.4,<3
pydantic==2.9.2
orjson==3.10.7
starlette==0.38.6
tenacity==9.0.0

